backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
google-generativeai
apscheduler
python-jose[cryptography]
passlib[bcrypt]
httpx
cryptography
newsapi-python
authlib
itsdangerous
